feat: publish README.md as index.html to S3 TDE-862 (#299)
### Description
Publish the `README.md` content as a `index.html` page to the
`s3://nz-imagery` S3 bucket.

![image](https://github.com/linz/imagery/assets/86932794/85a0e959-30a4-4cbf-acf2-b1ab8c23792b)



### Intention
When a user goes to https://nz-imagery.s3.ap-southeast-2.amazonaws.com,
the landing page is a XML document with the result of a `aws s3 list`
that is not very user friendly and can be confusing.

### Checklist
If not applicable, provide explanation of why.

- [ ] Tests updated: not critical
- [x] Docs updated
- [x] Issue linked in Title

---------

Co-authored-by: Paul Fouquet <paul@Pauls-MacBook-Air.local>