ci: remove parameters yaml filename (#377)
The logic to use `$base` (the name of the parameters.yaml file) to name
the argo workflow submitted no longer works.
https://github.com/linz/imagery/actions/runs/8413556393

This change will give each submission the name `publish-odr-`

In future we could update this to include the name of the branch/pull
request however as the action is currently broken this is the quickest
solution.