docs: revert to STAC Browser for browsing the catalog (#488)
### Motivation

STAC Index is using an older version of STAC Browser that isn't as easy
to browse through. We've received feedback that the previous link was
more helpful.

### Modifications

Reverted browsing links to go to STAC Browser rather than STAC Index.
Added a badge + link through to Registry of Open Data on AWS.