fix: remove marlborough_sn12080_sn122206_1994-1996_0.4m (#646)
### Motivation

There was a typo in the survey report so the wrong survey number was
used in the path and title when this was first published. We opted for a
re-standardisation and re-publish of the dataset to avoid manual fixing
the Collection (parent Catalog) and moving files across.
The correct dataset is to be published with
https://github.com/linz/imagery/pull/644

### Modification

- Delete STAC Collection

#### TODO

Delete files in
`s3://nz-imagery/marlborough/marlborough_sn12080_sn122206_1994-1996_0.4m/rgb/2193/`

### Verification

None
<!-- TODO: Say how you tested your changes. -->